--- packages/twenty-server/src/engine/api/rest/core-query-builder/utils/path-parsers/__tests__/parse-batch-path.utils.spec.ts
@@ -1,11 +1,0 @@
-import { parseBatchPath } from 'src/engine/api/rest/core-query-builder/utils/path-parsers/parse-batch-path.utils';
-
-describe('parseBatchPath', () => {
-  it('should parse object from request path', () => {
-    const request: any = { path: '/rest/batch/companies' };
-
-    expect(parseBatchPath(request)).toEqual({
-      object: 'companies',
-    });
-  });
-});
