+++ packages/twenty-server/src/engine/api/rest/core-query-builder/utils/path-parsers/parse-core-batch-path.utils.ts
@@ -1,5 +1,5 @@
 import { Request } from 'express';
 
-export const parseBatchPath = (request: Request): { object: string } => {
+export const parseCoreBatchPath = (request: Request): { object: string } => {
   return { object: request.path.replace('/rest/batch/', '') };
 };
