+++ packages/twenty-server/src/engine/api/rest/utils/clean-graphql-response.utils.ts
